Add the most recent SPF meet.
This commit marks a change in how meet results will be imported:
instead of getting all the results to fit a standard format,
they will just be converted to CSV format, with a header line
describing the column names.

Once there is a sufficient mass of data, the CSV files can be
parsed to standardize on column names, and to create files that
follow a certain common format.

The hard problem at the moment is just getting all the data
into a format that's easily able to be parsed, so we will
just solve that lesser problem first.