Remove Wilks where no BodyweightKg is provided.
Federations notoriously calculate Wilks improperly,
and often factor in a McCulloch age coefficient.

If there is insufficient information to regenerate
the Wilks score (BodyweightKg and TotalKg), then the
federation Wilks score is not to be believed,
and must be deleted. Lifters who competed in a weight
class will be assumed to be at the maximum bodyweight
of that class, except for SHWs, who will not receive
a Wilks score.