Use a ham-fisted binary search to improve compilation speed.
Building web/ moved from ~30s to ~2s, most of which is JS file compilation.
The binary search algorithm used is very stupid and requires that a result be present,
but it's sufficient for this use case. The `bisect` package unfortunately doesn't
take a key argument, so a lambda comparator can't be used.