Update lifters.csv
Changed Micheal Bingham's bench press result form Raw to Single-Ply.  If his result was raw not only would he have broken a world record but at the age of 49 he would have broken a 30 year old record by 50 lb, clearly this was miss labeled as Raw rather than Single Ply.
Also researching this man's name I could come up with no news of a WR or video of the lift so it's very unlikely that this was a Raw lift.