Update lifters.csv (#99)
* Update lifters.csv

In two different places on the score-sheet it shows Derrington Wright's bench press result as 193.5 and then 255 kg.  Either one of these entries is in error or the higher number was a 4th attempt in a bench shirt.  Both were labeled as Raw the only differentiation is one is under Open and the other is under Junior.  It's not uncommon for lifters to be entered into multiple weight classes, but the higher bench press is more than 120 lb higher than any of Derrington's other raw bench results so it appears to be either a equipped attempt or some kind of notation error.  Please advise.

* Update lifters.csv

* Update lifters.csv