Update lifters.csv
Ho boy, this one has problems.  From line 44-52 starting with Justin Haskins and ending with Andrew Olaya each lifter's bench result is the same as their deadlift result, and all their squat results are much lower.  What I fear happened here is where they were inputting the data at the meet they accidentally copied the bench and deadlift data over one row and overwrote the squat and bench data for those 9 lifters.  I can fix it so that it shows the data correctly in the bench and deadlift columns but there is no way for me to know what those lifters results in squat were.  Otherwise I guess we treat it as a push-pull?  What do you think the best course of action is here?