Fix locale issue in Mac OS X
There was an issue setting the locale on Mac OS X.
It turns out that Mac OS X only can handle
`en_US.UTF-8` and not `en_US.utf8`. So now if we fail
to set the locale to `en_US.utf8`, then we'll just go
ahead and try `en_US.UTF-8` as well.