Update entries.csv
Clearly this -198 lifter did not deadlift 1130lbs
This was the code in the results file:
<td>198</td>

    <td>AM</td>

    <td>Teen 17</td>

    <td>Dustin Pfeiffer</td>

    <td align="right">197.5</td>

    <td align="right">400</td>

    <td align="right">295</td>

    <td align="right">395</td>

    <td>1090</td>

  </tr>

  <tr>

    <td>&nbsp;</td>

    <td>&nbsp;</td>

    <td>&nbsp;</td>

    <td>&nbsp;</td>

    <td align="right">&nbsp;</td>

    <td align="right">4th</td>

    <td align="right">315</td>

    <td align="right">415</td>

    <td>1130</td>
From this I gleaned that the best attempts in 3 lifts were 400/295/395 for a total of 1090, the 4th lifts for bench and dead were 315/415.  They list a "4 lifts total" of 1130 but this is moot to us as we only recognize totals from 3 attempts.