Update social-media.csv
Added a bunch of instagrams from the top totals so far of 2018. Will DM them after they are added to the site and let them know. This used to be a great way to generate more followers and visits to the site for people that dont know about us so I'm trying to do it more again.