Force Instagram links to open in new tabs due to IG deleting browsing history.
Recently, Instagram has started making the "Back" button non-functional.
So if an OpenPowerlifting user is on a lifter page and clicks an IG link,
there is no way for them to return to that lifter page.

So since that's the behavior, it's better to just force IG links to open
in new tabs.