Take away Instagram's ability to modify our window's DOM.
Links opened with `target="_blank"` open in a new tab but are members
of the same tab group, with the newly opened tab apparently being able
to modify the window global of its parent tab.

Setting rel="noopener" prevents that behavior and allows IG/FB
to run in its own content process.