Version static assets. Closes #1353
Static assets are versioned manually (outside of WebPack) via
the helper script `server/client/apply-md5`, which operates
before gzipping.

The script inserts an MD5 hash into the filename and produces
a Tera template file as output. The template file defines each
file as a global variable, which can then be referenced at runtime
by the templating system.

While developing, the template is referenced via symlink;
when deploying, the symlink is overwritten with the file directly.