Allow AJAX requests to run to completion. Closes #1354
Creates the idea of a `pendingItem` work item, where an AJAX handler
always fetches the most recent work item, and can recursively
make further AJAX requests if the viewport updates during the request.

The `ensureData()` function just needs to make sure that a timeout is set
and that the `pendingItem` has the values of the most-recent viewport.