"Fix" all event warnings in the USAPL. 1833 left!
The USAPL data contains lots of errors created by the federation.
Instead of marking lifters DQ where the federation gave them a placing,
we just do whatever the federation did, and accept that as correct.

This does lead to a weird situation where "SD" lifters compete
directly against "SBD" and "BD" lifters, but that's what the federation
seems to have recorded in their records.