Fix incorrect attempt and placing data in usapl-archive/NS-2003-08-22. Closes #4065
The USAPL appears to have marked every first attempt as a failure.

It was fixed by taking the placing/DQ information from the URL manually,
marking each first attempt as a success, and then fixing errors until everything
reconciled with the checker. Some failed firsts may be marked as successes,
but only in cases where it doesn't affect the outcome.