Show only meets from the exact federation on the Meet List page. Closes #4366
The URL segment "uspa" can be parsed as either Federation::USPA or as
MetaFederation::USPA. Usually, for records and rankings, we prefer the latter.

The Meet List page is different: if we parse as a MetaFederation, all the
international affiliate meets show up too, which makes it hard to find the meets
from a single federation.

This patch introduces a switch in the segment parser that tells it to try
parsing as a Federation first. The Meet List page then uses that switch.