Revert "The meet entered as fpo/2001 was unsanctioned per Kalle. Removing. #5348"
This reverts commit 1fd2ae2bce40202a322cbbf82aa983ac5a2a9e6d.

The FPO sanction was added retroactively due to confusion around the
meet's sanction. The meet's host, Anna Khudayarov, believed that *WE*
had green-lit including her meet even without a sanction, so she told
Kalle that the FPO didn't have to get involved. We never intended to do
that. Kalle agreed to sanction the meet retroactively to avoid
confusion, since everyone believed it was a real meet and treated it as
such.