Significantly improve set intersection throughput
This patch changes NonSortedNonUnique::intersect() to use iterators
instead of vectors. As a result, bounds checks are elided. It turns out
that the vast majority of the time was just spend checking bounds.

A benchmark shows that calculating the intersection of Raw+Wraps and Male
moves from 6.4ms to 4.5ms.