Add a global DISTRIBUTION and use the IS_MOBILE global
We previously had mobile base.tera files set an IS_MOBILE global
variable, but the isMobile() function looked for footer presence instead
of just respecting the global. Because the new isMobile() no longer
requires DOM interaction, it's cheap to call, so we do it more
frequently.

This also fixes a bug on the FAQ page caused by the last code change,
where the undefined case of PAGE_KIND was not handled properly.