feat: add lifter Instagram script
Adding lifter Instagram's is a fairly easy process, just requiring you
to add a single line in the right place inside
`lifter-data/social-instagram`. This can easily be automated, by simply
loading the file, iterating to find where we should add the entry and
then writing back to the file.

Usage is as follows:
```
./scripts/add-lifter-instagram --name "<name>" --instagram "<handle>"
```

From there, you can simply add the change and make a pull request. This
was used for !7058, so it is at least somewhat functional.