Disable CI on the main branch due to GitLab runner limits
GitLab sent us a message letting us know that we're at risk of running
out of shared runtime minutes, which is a new thing they're starting to
bill for this year. I attempted to get us registered as an open-source
project, which they approved, but I wasn't able to apply that setting to
our projects.

We don't really care about most of the CI on main, except for the pages
publishing, which are almost always unchanged. However, we do want CI in
MRs, because our team uses that. This patches things up for the moment.