Add Daniel Austin 1 as a Raw Deadlifter in plusa/1992-09-03-A for ATWR
Daniel Austin competed Single-ply but deadlifted Raw. That deadlift is
still the raw deadlift record to this day.

Our system isn't capable of handling equipment changes during the course
of the competition, but we want to show this record. To work around
this, we're adding an extra Guest entry to this meet with just the Raw
record we want to display.