feat: add email templates
In #7817 we moved all the "Contact" information to the same template,
such that it could be shared across mobile, desktop, OpenPowerlifting
and OpenIPF without being duplicated for each of them.

Now, we can add some email template links. Initially this only adds
templates for Instagram/DoB queries as well as name corrections, since
other queries are a bit harder to define a template for and it would be
good to see these being used first.

This change:
* Adds a dependency on `urlencoding` so that we can, well, URL encode
  things for people to click on
* Adds functions to generate the 2 new templates
* Adds them to the `Context` struct for the contact page
* Updates the copy in the template to have a "Common Issues" section to
  hopefully guide people towards the email templates