ci: enable timings output
CI seems to be taking a lot of time (~12 minutes for a PR) so it would
be good to try and cut that down, both for faster feedback but also to
use less minutes of the allowance.

The largest timesave will likely be the removal of compilation for
`checker` itself, since it rarely changes and thus there's no real need
to build it every time. This should save at least 3 minutes (according
to the output of `cargo`), but there seems to be a lot of other time
spent.

Would be good to get an output of each of the phases for a PR when doing
the checking itself. Locally, my laptop takes ~3 seconds for a full
check which is pretty good, but the GitLab CI agents are probably quite
a bit slower.

This change:
* Updates the labels for timings to be more descriptive
* Enables `--timing` by default for CI