feat: validate filenames in meet directories
In order to standardise the structure of the repository a little more
and provide better errors when meets fail to process, we can validate
that the meet directories contain a specific set of filenames. This will
catch errors such as `Entries.csv` or other issues.

This is relatively relaxed for the time being, largely just picking the
set of files that already exist until the error count is low enough to
manually fix the rest.

This change:
* Adds some validation to the `checker` to ensure meet directories
  contain sensibly named files