Allow specifying "Bare" and "Sleeves" in the Equipment column
Newer versions of OpenLifter will begin exporting the Equipment column
exactly as specified during the meet, in which Bare and Sleeves exist as
values, and Raw does not, since the term is massively overloaded.

For purposes of OpenPowerlifting, Bare and Sleeves are treated as
aliases of Raw, and become Raw during CSV parsing.

The goal of this change is to lower importation difficulty, not to
create separate Bare and Sleeves categories in OpenPowerlifting.