fix(checker): Allow .DS_Store files in meet directories
macOS apparently creates a lot of these files. They're already ignored
in the .gitignore, so showing errors for them creates a lot of busywork
to no benefit. This patch marks them as a valid entry (since we are
aware of them, and can ignore them).

Note that we don't want to parse our .gitignore and allow everything in
there in the general case, since the GitLab WebUI bypasses it and still
allows committing XLSX files, which is how we got here in the first
place.