perf(db): Reduce memory usage by 11% using symbol tables
Using GlobalSymbol instead of CompactString uses less memory for storing
Division by moving those strings out of the Entry struct. We therefore
get better locality for Entry rows.

The only downside of this patch is that the GlobalTable doesn't support
accurate memory accounting, so you have to look at RSS of the server as
it's up, which is somewhat misleading. We could write our own symbol
table implementation in the future to allow for accurate tracking.