Update DQ to G in uspa/2270. #20432
This is some editorialization on my part based on the description of
what occurred in the federation in the issue. The federation was
inconsistent in applying its own rules, but does not want the lifts to
count toward a record category. They disqualified the lifter, but could
have accomplished the same thing using the Guest category. I imagine
this did not occur to them. This remedy is superior and does not change
USPA rankings.