fix(checker): Make ExemptDuplicates independent of date sort order. Closes #20887
Prior to this patch, if two meets were duplicates of each other, it was
possible to pass the checker non-deterministically if ExemptDuplicates
is set for only one of the meets in the pair.

If the meet with ExemptDuplicates sorted prior to the other meet, it
would pass. If the meet sorted after, it would fail. The order of the
meets was determined by the order in which the checker visited them.

This patch fixes the issue by checking ExemptDuplicates for both meets.
That means that it's legal to only set it for a single meet in a pair,
and the checker will then always pass.