feat(checker): Only check entries.csv if meet.csv parses
New contributors using our CI system are not used to reading error
messages starting from the first error emitted, and begin from the last
error omitted.

This causes a problem if the `meet.csv` file fails parsing. When that
happens, default values are used for several configuration options. For
configured federations with date-based logic, that means that the
configuration can be ignored, and the checker will emit false positives,
typically about weightclasses. The contributor then gets confused, and
we have to explain that the error really lies with the meet.csv.

This patch skips that whole cycle by refusing to parse the entries.csv
if the meet.csv isn't already correct. Then the last error shown will be
the actual error, and the contributor can self-help.

Interestingly, the Online Checker hosted by the server already behaved
this way.