perf(db): Eliminate overhead on lifter page render
A profile on the server showed that the majority of CPU time is spent in
the function `opldb::lifters_under_username()`, which is called on
lifter page render. Although user pages render quickly (~10ms), they
consume a significant amount of resources because there is a very large
number of lifters, and multiple web crawlers insist on visiting them
all, often repeatedly.

That function attempts to determine whether the provided string is a
disambiguation or not, in order to determine what kind of template to
render. To do this, the previous version iterated over every lifter.
That is wasteful, because a perfectly good hashmap exists to perform
O(1) point lookups on the small amount of username possibilities.

This patch adds a benchmark and changes the implementation to use the
already-existing hashmap. Execution time in the benchmark improves from
8.19ms (84_568ns) to 97ns, which is 0.0011% of the previous execution
time.

This eliminates the majority of server load.