perf(db): Eliminate overhead of meet page render
After fixing the lifter page performance, another server profiled that
about 60% of compute time was now spent calculating entries for
individual meet pages. Although this operation is still relatively fast
(~10ms), because of the large amount of web crawler traffic asking for
meet pages, it occupies the majority of CPU time.

Once again, the previous algorithm was searching the backing array, and
the solution is to use a simple cache. For this patch, that cache is an
array of entry_ids sorted by meet_id. Constructing that array adds about
10MB to resident memory and about 0.5s to startup time (16.5 -> 17s),
which are both negligible. In exchange, the entries-for-meet lookup
operation takes ~700ns instead of ~10ms (10_000_000ns).

This again eliminates the majority of server load. Making the lifter
pages more efficient reduced server CPU load from ~40% to ~3.3%. This
patch should reduce it to <1.0%.