server: Link to New Issue on Gitlab instead of the Contact page
Gmail's spam filter is again way too aggressive, and too much is getting
lost. This is an experiment to see if we can stop recommending email
from this point forward, and instead point people directly to our New
Issues page.

In practice, this will take them to the sign-in page, where they can log
in with a Google account or create another one manually. It shouldn't be
too bad. Email will of course still work for those who can get past the
spam filter.