Delete Bill Helmich "memorial" lift from amp/2025-NS-05
Mike Z set up a memorial tribute and put Bill into Liftingcast with the rest of the Masters with his existing world record deadlift as the lift. This allowed us to white light him 1 final time. I just forgot to take it out before uploading him here and it would have caused problems on the rankings if I left him in.