checker: Support 'options.valid_federations' in CONFIG.toml
You can now add an 'options.valid_federations' section to a CONFIG.toml
file, which restricts the range of federations accepted under that
config file.

For example, here is a sample USAPL configuration:

```toml
 [options]
     valid_federations = ["USAPL", "NAPF"]
     valid_since = "2017-07-01"
     require_manual_disambiguation = true
```

This change serves two purposes:

 1. For federations like WRPF-AUS that often submit data under their
    parent federation, WRPF, this is a way to automatically assert that
    the federation uploaded matches expectation, without us having to
    catch that error manually.

 2. For the openpowerlifting.org/dev checker, this provides us with a
    hint as to which CONFIG.toml files correspond to which federations,
    letting us take a step closer to supporting config files in that
    environment.

The 'options.valid_federations' field is optional for now, though it is
strongly suggested to be used. Once used pervasively, we can make it
mandatory. Maybe we can even assert that configurations are disjoint!